Move auth.py to service layer
